+++ src/main/kotlin/us/frollo/frollosdk/extensions/APIExtension.kt
@@ -329,7 +329,8 @@ internal fun CdrAPI.fetchConsents(
     providerAccountId: Long? = null,
     after: Long? = null,
     before: Long? = null,
-    size: Long? = null
+    size: Long? = null,
+    cdrConfigExternalId: String
 ): Call<PaginatedResponse<ConsentResponse>> {
     val queryMap = mutableMapOf<String, String>()
     status?.let { queryMap.put("status", status.toString()) }
@@ -338,6 +339,7 @@ internal fun CdrAPI.fetchConsents(
     after?.let { queryMap.put("after", it.toString()) }
     before?.let { queryMap.put("before", it.toString()) }
     size?.let { queryMap.put("size", it.toString()) }
+    cdrConfigExternalId.let { queryMap.put("cdrConfigExternalId", it) }
     return fetchConsents(queryMap)
 }